pyproject.toml
[tool.poetry]
name = "drf-courses"
version = "0.1.0"
description = ""
authors = ["Aratimius <aratimius@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.1"
psycopg2-binary = "^2.9.9"
pillow = "^10.4.0"
djangorestframework = "^3.15.2"
ipython = "^8.28.0"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.8"
python-dotenv = "^1.0.1"
stripe = "^11.2.0"
celery = "^5.4.0"
redis = "^5.2.0"
django-celery-beat = "^2.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
